Bug fix in helper functin getRMS()
The calculation of the RMS was missing a factor of 0.5 (within the sqrt).

This wasn't a problem in the original code because the calculation of the wave amplitudes also missed that factor, so they cancelled out. But I forgot to add this factor in getRMS() when we fixed the calculation of the wave amplitude